RELEASE v2.2.0

+++ src/ssh-setup.js
@@ -1,7 +1,8 @@
 const { execFileSync, execSync } = require('child_process')
 const fs = require('fs')
 const {
-  info
+  info,
+  exportVariable
 } = require('@actions/core')
 
 const { convertActionToCloneCommand } = require('./action-parser')
@@ -13,7 +14,7 @@ const sshHomeSetup = () => {
   execSync(`ssh-keyscan -H github.com >> ${sshHomePath}/known_hosts`)
 }
 
-const sshAgentStart = () => {
+const sshAgentStart = (exportEnv) => {
   info('SSH > Starting the SSH agent')
   const sshAgentOutput = execFileSync('ssh-agent')
   const lines = sshAgentOutput.toString().split('\n')
@@ -21,6 +22,10 @@ const sshAgentStart = () => {
     const matches = /^(SSH_AUTH_SOCK|SSH_AGENT_PID)=(.*); export \1/.exec(lines[lineNumber])
     if (matches && matches.length > 0) {
       process.env[matches[1]] = matches[2]
+      if (exportEnv) {
+        exportVariable(matches[1], matches[2])
+        info(`SSH > Export ${matches[1]} = ${matches[2]}`)
+      }
     }
   }
 }
@@ -33,10 +38,16 @@ const addPrivateKey = (privateKey) => {
   execSync('ssh-add -l', { stdio: 'inherit' })
 }
 
-const sshSetup = (privateKey) => {
+const sshSetup = (privateKey, exportEnv) => {
   sshHomeSetup()
-  sshAgentStart()
+  sshAgentStart(exportEnv)
   addPrivateKey(privateKey)
+}
+
+const configureSSHGit = () => {
+  const command = 'git config --global url."ssh://git@github.com/".insteadOf "https://github.com/"'
+  info(`App > ${command}`)
+  execSync(command)
 }
 
 const cloneWithSSH = (basePath, action) => {
@@ -48,13 +59,16 @@ const cloneWithSSH = (basePath, action) => {
 }
 
 const cleanupSSH = () => {
-  info('SSH > Killing the ssh-agent')
-  /* eslint no-template-curly-in-string: "off" */
-  execSync('kill ${SSH_AGENT_PID}', { stdio: 'inherit' })
+  if (process.env.SSH_AGENT_PID) {
+    info('SSH > Killing the ssh-agent')
+    /* eslint no-template-curly-in-string: "off" */
+    execSync('kill ${SSH_AGENT_PID}', { stdio: 'inherit' })
+  }
 }
 
 module.exports = {
   sshSetup,
   cloneWithSSH,
-  cleanupSSH
+  cleanupSSH,
+  configureSSHGit
 }
